Update fn_initQRFSystemForPreplacedGroups.sqf
Added a short delay (`sleep 0.1`) between the check of each unit and each group. This will take way more time but it gives some room for the rest of the functions and commands to run. It was mostly added in case the function will be called in run-time and not on mission start!

Co-Authored-By: A. Ares <arismagicgr@gmail.com>