Update fn_spawnQRF.sqf
Optimized the code a bit more by changing the calculation of the nearest object from using BIS_fnc_sortBy with ascending sorting and taking the first (nearest) element of the sorted array, with the BIS_fnc_nearestPosition function which actually does exactly what it is needed here.

Also, deleted a local variable that was not really needed anymore!